Add working push_swap.. bad Makefile and uses too many moves though (also norm broken)

+++ libft/ft_atoi.c
@@ -10,9 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_atoi(const char *str)
+long		ft_atoi(const char *str)
 {
-	int num;
+	long num;
 	int i;
 	int neg;
